LibWeb: Move WebDriver's HeapTimer helper class to its own file
And generalize it a tiny bit to be reusable outside of ExecuteScript.

(cherry picked from commit 8598d4670dc6a5b44f2081457af10ea5b2306108)
