+++ Userland/Libraries/LibWeb/Forward.h
@@ -816,6 +816,8 @@ using Promise = JS::PromiseCapability;
 }
 
 namespace Web::WebDriver {
+class HeapTimer;
+
 struct ActionObject;
 struct InputState;
 };
